% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRC_combine.R, R/updated_CRC_combine.R
\name{combine_by_age_cat_sex}
\alias{combine_by_age_cat_sex}
\title{Combine MCED and CRC data for specific age category and sex}
\usage{
combine_by_age_cat_sex(
  the_CRC_data,
  combined_additional_results,
  combined_first_results,
  the_age_cat
)

combine_by_age_cat_sex(
  the_CRC_data,
  combined_additional_results,
  combined_first_results,
  the_age_cat
)
}
\arguments{
\item{the_CRC_data}{Processed CRC data with age categories}

\item{combined_additional_results}{MCED additional cancer outcomes}

\item{combined_first_results}{MCED first cancer diagnoses}

\item{the_age_cat}{Character string: age category (e.g., "(45,50]")}
}
\value{
A list with with results_all: Combined MCED + CRC outcomes for age category
no_match_all: Total number of unmatched reassignments summed
over males and females

A list with with results_all: Combined MCED + CRC outcomes for age category
no_match_all: Total number of unmatched reassignments summed
over males and females
}
\description{
Integrates CRC screening outcomes with MCED results within a specified 5-year
other-cause-death age category, processing males and females separately and
then combining the outputs.

Integrates CRC screening outcomes with MCED results within a specified 5-year
other-cause-death age category, processing males and females separately and
then combining the outputs.
}
\examples{
out_age_bin <- combine_by_age_cat_sex(
                           the_CRC_data = crc_processed,
            combined_additional_results = combined_addtl,
                 combined_first_results = combined_first,
                            the_age_cat = "(60,65]")
                             
out_age_bin <- combine_by_age_cat_sex(
                           the_CRC_data = crc_processed,
            combined_additional_results = combined_addtl,
                 combined_first_results = combined_first,
                            the_age_cat = "(60,65]")

}
