% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicancer_sim.R
\name{create_emission_matrix}
\alias{create_emission_matrix}
\title{Create emission matrix for a given rate matrix and sensitivities}
\usage{
create_emission_matrix(
  rate_matrix,
  early_sensitivity,
  late_sensitivity,
  specificity
)
}
\arguments{
\item{rate_matrix}{Numeric matrix. Rate matrix for the cancer site.}

\item{early_sensitivity}{Numeric. Early sensitivity for the cancer site.}

\item{late_sensitivity}{Numeric. Late sensitivity for the cancer site.}

\item{specificities}{Numeric. Specificities for each cancer site.}
}
\value{
A numeric matrix representing the emission matrix.
}
\description{
This function creates an emission matrix for a given rate matrix and sensitivities.
}
