% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.ctmc.R
\name{discrete.ctmc}
\alias{discrete.ctmc}
\title{Discretize a continuous-time Markov chain (CTMC) trajectory}
\usage{
discrete.ctmc(ctmc.times, ctmc.states, obs.times)
}
\arguments{
\item{ctmc.times}{Transition times of the CTMC}

\item{ctmc.states}{States of the CTMC at corresponding transition times.}

\item{obs.times.}{Times at which observations are desired.}
}
\value{
A data frame containing the oberved times and corresponding states.
}
\description{
This function discretizes a CTMC trajectory given the transition times and states.
}
\examples{
# Discretize a CTMC trajectory
times <- c(0, 1, 2, 3, 4)
states <- c(1, 2, 1, 2, 1)
obs_times <- c(0.5, 1.5, 2.5, 3.5)
discretized_data <- discrete.ctmc(ctmc.times = times, ctmc.states = states, obs.times = obs_times)

}
