% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cause_death.R
\name{filter_multiple_cancers}
\alias{filter_multiple_cancers}
\title{Filter MCED cancer data for selected cancer types}
\usage{
filter_multiple_cancers(data, cancer_sites)
}
\arguments{
\item{data}{A data frame containing MCED cancer mortality data. This data frame should
include a column named \verb{ICD-10 113 Cause List} which specifies the cancer types.}

\item{cancer_sites}{A vector of selected cancer type names to filter. These names
should match the values in the \verb{ICD-10 113 Cause List} column of the data frame.}
}
\value{
A data frame containing only rows where cancer type is in \code{cancer_sites}.
}
\description{
This function filters the input MCED cancer data to include only the cancer types
specified in the \code{cancer_sites} vector. It uses the \code{filter} function from \code{dplyr}
to select rows where the cancer type matches any of the types listed in \code{cancer_sites}.
}
\examples{
selected_cancers <- c(
"Malignant neoplasm of prostate (C61)",
"Malignant neoplasm of bladder (C67)"
)
filtered_data <- filter_multiple_cancers(MCED_cdc, selected_cancers)
}
