% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.ctmc.R
\name{get.obs.data.individual.control}
\alias{get.obs.data.individual.control}
\title{Get observed data for an individual from of an individual in the control arm based on the CTMC natural history trajectory}
\usage{
get.obs.data.individual.control(
  ID,
  rate.matrix,
  end.time = 30,
  start.time = 0,
  start.state = 1
)
}
\arguments{
\item{ID}{The identifier for the individual.}

\item{rate.matrix}{The rate matrix of the CTMC.}

\item{end.time}{The end time for the CTMC trajectory (default is 30).}

\item{start.time}{The start time for the CTMC trajectory (default is 0).}

\item{start.state}{The start state for the CTMC trajectory (default is 1).}
}
\value{
A data frame containing the identifier, screen diagnosis time, screen diagnosis stage, clinical diagnosis time, and clinical diagnosis stage.
}
\description{
This function generates observed data for an individual from a continuous-time Markov chain (CTMC) trajectory and Hidden Markov Model (HMM).
}
\examples{
# Get observed data for an individual
rate_matrix <- matrix(c(-0.1, 0.1, 0.2, -0.2), nrow = 2)
obs_data_individual <- gets.obs.data.individual.control(ID = 1, rate.matrix = rate_matrix, end.time=30,start.time=0,start.state=1)

}
