% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.ctmc.R
\name{get.obs.data.many}
\alias{get.obs.data.many}
\title{Generate observed data for multiple individuals from a CTMC trajectory and Hidden Markov Model (HMM)}
\usage{
get.obs.data.many(
  num.individuals,
  rate.matrix,
  emission.matrix,
  obs.times = seq(1, 30, 2),
  end.time = 30,
  start.time = 0,
  start.state = 1
)
}
\arguments{
\item{num.individuals}{The number of individuals for which to generate observed data.}

\item{rate.matrix}{The rate matrix of the CTMC.}

\item{emission.matrix}{The emission matrix of the HMM.}

\item{obs.times}{The observation times (default is seq(1, 30, 2)).}

\item{end.time}{The end time for the CTMC trajectory (default is 30).}

\item{start.time}{The start time for the CTMC trajectory (default is 0).}

\item{start.state}{The start state for the CTMC trajectory (default is 1).}
}
\value{
A data frame containing the observed data for multiple individuals.
}
\description{
This function generates observed data for multiple individuals from a continuous-time Markov chain (CTMC) trajectory and Hidden Markov Model (HMM).
}
\examples{
# Generate observed data for multiple individuals
rate_matrix <- matrix(c(-0.1, 0.1, 0.2, -0.2), nrow = 2)
emission_matrix <- matrix(c(0.2, 0.3, 0.5, 0.1, 0.4, 0.5), nrow = 2)
observed_data_many <- get.obs.data.many(num.individuals = 10, rate.matrix = rate_matrix, emission.matrix = emission_matrix)

}
