% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MCED_helper_fun.R
\name{get_filtered_rates}
\alias{get_filtered_rates}
\title{Extract rate matrices and cancer sites based on OMST and LMST specifications.}
\usage{
get_filtered_rates(
  the_omsts,
  the_lmsts,
  all_meta_data,
  all_rates,
  the_cancer_sites
)
}
\arguments{
\item{the_omsts}{vector of overall mean sojourn times for each of the cancer specified cancer sites}

\item{the_lmsts}{vector of late mean sojourn times for each of the cancer specified cancer sites}

\item{all_meta_data}{dataframe that indexes the rate matrices corresponding to specified OMST and LMST entries}

\item{all_rates}{array of rate matrices corresponding to specified OMST and LMST entries}
}
\value{
A list with two items:
1) A list of rate matrices that correspond to the selected OMST and LMST specifications for each the selected cancer sites.
2) vector of the cancer sites
}
\description{
Extract rate matrices and cancer sites based on OMST and LMST specifications.
}
