% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_MCED_helper_fun.R
\name{get_init}
\alias{get_init}
\title{Get the initial natural history state based on the rate matrix and starting age, conditional on no clinical diagnoses before starting age.}
\usage{
get_init(rate.matrix, a1)
}
\arguments{
\item{rate.matrix}{the rate matrix used to simulate cancer natural history}

\item{a1}{starting age at first screen}
}
\value{
the initial state
}
\description{
Get the initial natural history state based on the rate matrix and starting age, conditional on no clinical diagnoses before starting age.
}
