% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cause_death.R
\name{get_other_cause_mortality}
\alias{get_other_cause_mortality}
\title{Adjust all-cause mortality to estimate other-cause mortality rates}
\usage{
get_other_cause_mortality(cdc_data, MCED_cdc, hmd_data, selected_cancers)
}
\arguments{
\item{cdc_data}{A data frame with all-cause mortality data from the CDC. This data frame should
include columns: "Single-Year Ages Code", "Gender", "Year", "Deaths", "Population", "Crude Rate".}

\item{MCED_cdc}{A data frame with cancer-specific mortality data from the CDC. This data frame should
include columns: "Single-Year Ages Code", "Sex", "Year", "ICD-10 113 Cause List",
"ICD-10 113 Cause List Code", "Deaths", "Population", "Crude Rate".}

\item{hmd_data}{A data frame with population data from HMD. This data frame should include columns:
"Year", "Age", "Female", "Male", "Total".}

\item{selected_cancers}{A character vector specifying the cancer types to analyze.
These names include: "Bladder","Breast","Colorectal", "Esophagus","Gastric", "Headandneck",
"Liver", "Lung","Ovary", "Pancreas", "Prostate", "Renal", "Uterine"}
}
\value{
A data frame with age-specific and sex-specific mortality rates due to causes
other than the selected cancers. Includes columns: age, sex, year, other_cause_rate.
}
\description{
This function calculates mortality rates for causes other than the selected (MCED) cancers.
All cause and cause-specific (MCED) death rates for 2018-2022 were obtained from CDC (Centers
for Disease Control and Prevention 2025) and HMD (Human Mortality Database 2025) by age and gender.
}
\details{
The function performs several steps:
\enumerate{
\item Filters the MCED cancer data for the selected cancers.
\item Converts HMD data to a long format for easier processing.
\item Process all cause CDC and HMD data to generate all cause death rates
\item Convert all-cause death rates to probabilities using formula of Rosenberg (2006).
\item Process MCED CDC and HMD cancer data to generate MCED cancer death rates
\item Convert MCED cancer death rates to probabilities using formula of Rosenberg (2006).
\item Compute probability of death from causes other than MCED cancers
\item Convert other cause death probabilities to rates using inverse formula of Rosenberg (2006).
}
}
\examples{
cdc_data <- read.csv("all_cause_cdc.csv")
MCED_cdc <- read.csv("MCED_cdc.csv")
hmd_data <- read.csv("hmd_data.csv")
selected_cancers <- c(
"Prostate",
"Bladder"
)
get_other_cause_mortality <- adjusted_all_cause_mortality(cdc_data, MCED_cdc, hmd_data, selected_cancers)
}
