% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_death.R
\name{gettime}
\alias{gettime}
\title{Simulate a time based on a provided survival function.}
\usage{
gettime(time, surv)
}
\arguments{
\item{time}{vector of times corresponding to the survival function}

\item{surv}{vector of survival probabilities at each time.}
}
\value{
A data frame with "time" and "status" elements, where status=1 indicates the event occurred.
}
\description{
Simulate a time based on a provided survival function.
}
