% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cause_death.R
\name{make_othercause_death_table}
\alias{make_othercause_death_table}
\title{Create a table of other-cause death survival probabilities}
\usage{
make_othercause_death_table(
  cdc_data,
  MCED_cdc,
  hmd_data,
  selected_cancers,
  the_sex,
  the_year,
  the_starting_age
)
}
\arguments{
\item{cdc_data}{A data frame with all-cause mortality data from the CDC. This data frame should
include columns: "Single-Year Ages Code", "Gender", "Year", "Deaths", "Population",
"Crude Rate".}

\item{MCED_cdc}{A data frame with cancer-specific mortality data from the CDC. This data frame should
include columns: "Single-Year Ages Code", "Sex", "Year", "ICD-10 113 Cause List",
"ICD-10 113 Cause List Code", "Deaths", "Population", "Crude Rate".}

\item{hmd_data}{A data frame with population data from HMD. This data frame should include columns:
"Year", "Age", "Female", "Male", "Total".}

\item{selected_cancers}{A character vector specifying the cancer types to analyze.
These names include: "Anus", "Bladder", "Breast", "Colorectal", "Esophagus", "Gastric",
"Headandneck", "Liver", "Lung", "Ovary", "Pancreas", "Prostate", "Renal", "Uterine".}

\item{the_sex}{A character string specifying the sex to filter by ("Male" or "Female").}

\item{the_year}{An integer specoifying the year to filter by.}

\item{the_starting_age}{An integer specifying the starting age for the survival probabilities.}
}
\value{
A data frame with survival probabilities for causes other than the selected cancers.
Includes columns: age, sex, year, other_cause_rate, surv.
}
\description{
This function generates a table of survival probabilities for causes other than selected (MCED) cancers.
Its uses the output from the \code{get_other_cause_mortality} function and filters it based on the specified
sex, year, and starting age.
}
\examples{
cdc_data <- read_excel("/path/to/cdc_all_cause_2018_2022.xlsx")
MCED_cdc <- read_excel("/path/to/modified_MCED_data.xlsx")
hmd_data <- read_excel("/path/to/hmd_population_1933_2023.xlsx")
selected_cancers <- c("Anus", "Bladder", "Breast", "Colorectal", "Esophagus", "Gastric", "Headandneck",
                      "Liver", "Lung", "Ovary", "Pancreas", "Prostate", "Renal", "Uterine")
othercause_surv <- make_othercause_death_table(cdc_data, MCED_cdc, hmd_data, selected_cancers, "Female", 2019, 60)
}
