% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.ctmc.R
\name{observed.data.hmm}
\alias{observed.data.hmm}
\title{Generate observed data from a Hidden Markov Model (HMM)}
\usage{
observed.data.hmm(obs.times, underlying.states, emission.matrix)
}
\arguments{
\item{obs.times}{The observation times.}

\item{underlying.states}{The underlying states of the HMM.}

\item{emission.matrix}{The emission matrix representing the probabilities of observing each state given an underlying state.}
}
\value{
A data frame containing the observation times and corresponding observed data points.
}
\description{
This function generates observed data from a Hidden Markov Model (HMM) given the observation times, underlying states, and emission matrix.
}
\examples{
# Generate observed data from an HMM
obs_times <- c(0.5, 1.5, 2.5, 3.5)
underlying_states <- c(1, 2, 1, 2)
emission_matrix <- matrix(c(0.2, 0.3, 0.5, 0.1, 0.4, 0.5), nrow = 2)
observed_data <- observed.data.hmm(obs.times = obs_times, underlying.states = underlying_states, emission.matrix = emission_matrix)

}
