% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_death.R
\name{sim_cancer_death}
\alias{sim_cancer_death}
\title{Simulate a time of cancer death for a specific cancer
Simulate a time of cancer death for a specific cancer in a specific stage based on provided survival distributions.
Uses the inverse CDF method.}
\usage{
sim_cancer_death(
  the_stage,
  the_cancer_site,
  the_sex,
  the_model_type,
  cancer_survival_dist
)
}
\arguments{
\item{the_stage}{stage at clinical diagnosis}

\item{the_cancer_site}{cancer site (one of "Anus","Breast","Bladder","Colorectal","Esophagus","Headandneck","Gastric","Liver" ,"Lung","Pancreas", "Prostate", "Renal", "Ovary", "Uterine")}

\item{the_sex}{Male or Female}

\item{the_model_type}{Weibull or  Loglogistic}

\item{cancer_survival_dist:}{a data frame with columns: surv, time, site, stage, sex, model_type}
}
\value{
A data frame with "time" and "status" elements, where status=1 indicates the event occurred.
}
\description{
Simulate a time of cancer death for a specific cancer
Simulate a time of cancer death for a specific cancer in a specific stage based on provided survival distributions.
Uses the inverse CDF method.
}
