% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_death.R
\name{sim_cancer_death_param}
\alias{sim_cancer_death_param}
\title{Simulate a time of cancer death for a specific cancer in a specific stage based on provided parametric survival distributions.}
\usage{
sim_cancer_death_param(
  the_stage,
  the_cancer_site,
  the_sex,
  ID = NA,
  the_model_type,
  param_table
)
}
\arguments{
\item{the_stage}{stage at clinical diagnosis}

\item{the_cancer_site}{cancer site (one of "Anus","Breast","Bladder","Colorectal","Esophagus","Headandneck","Gastric","Liver" ,"Lung","Pancreas", "Prostate", "Renal", "Ovary", "Uterine")}

\item{the_sex}{Male or Female}

\item{ID}{optional ID to set seed}

\item{the_model_type}{Weibull or  Loglogistic}

\item{param_table}{a data frame with columns: intercept (based on survreg fit), scale (based on survreg fit) site, stage, sex, model_type}
}
\value{
A data frame with "time" and "status" elements, where status=1 indicates the event occurred.
}
\description{
Simulate a time of cancer death for a specific cancer in a specific stage based on provided parametric survival distributions.
}
