% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicancer_sim.R
\name{sim_multiple_cancer_multiple_individuals}
\alias{sim_multiple_cancer_multiple_individuals}
\title{Simulate multiple cancer sites for multiple individuals}
\usage{
sim_multiple_cancer_multiple_individuals(
  num_individuals,
  cancer_sites,
  rate_matrices,
  early_sensitivities,
  late_sensitivities,
  specificities,
  obs.times,
  end.time,
  start.time,
  start.state
)
}
\arguments{
\item{num_individuals}{Number of individuals to simulate.}

\item{cancer_sites}{Character vector specifying the names of cancer sites.}

\item{rate_matrices}{List of rate matrices corresponding to each cancer site.}

\item{early_sensitivities}{Numeric vector of early sensitivities for each cancer site.}

\item{late_sensitivities}{Numeric vector of late sensitivities for each cancer site.}

\item{specificities}{Numeric vector of specificities for each cancer site.}

\item{obs.times}{Numeric vector specifying observation times.}

\item{end.time}{Numeric value specifying the end time of the simulation.}

\item{start.time}{Numeric value specifying the start time of the simulation.}

\item{start.state}{Numeric value specifying the initial state of the simulation.}
}
\value{
A data frame containing simulated observations for each cancer site across multiple individuals.
}
\description{
This function simulates the progression of multiple cancer sites for multiple individuals over a specified time period.
}
\examples{
sim_multiple_cancer_multiple_individuals(num_individuals = 50, cancer_sites = c("Lung", "Liver"),
                                          rate_matrices = list(lungrate, liverate),
                                          early_sensitivities = c(0.3, 0.2),
                                          late_sensitivities = c(0.9, 0.8),
                                           specificities = c(0.9, 0.85),
                                          obs.times = c(0, seq(50, 75, by = 2)), end.time = 100,
                                          start.time = 0,
                                          start.state = 1)

}
