% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_cause_death.R
\name{sim_othercause_death}
\alias{sim_othercause_death}
\title{Simulate other-cause death time}
\usage{
sim_othercause_death(othercause_death_table, ID = NA)
}
\arguments{
\item{othercause_death_table}{A data frame with survival probabilities for causes other than the selected cancers.
Includes columns: age, sex, year, other_cause_rate, surv.}

\item{ID}{optional ID to set random seed.}
}
\value{
A numeric value representing the simulated time to death from other causes.
}
\description{
This function simulates the time to death from causes other than the selected (MCED) cancers
using the survival probabilities from the \code{make_othercause_death_table} function.
}
\examples{
othercause_death_table <- make_othercause_death_table(cdc_data, MCED_cdc, hmd_data, selected_cancers, "Female", 2019, 60)
sim_time <- sim_othercause_death(othercause_death_table)
}
